place_remeber/requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.2
django-extensions==3.2.3
django-leaflet==0.30.0
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
django-sslserver==0.22
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
isodate==0.6.1
jwcrypto==1.5.6
lxml==5.2.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pep8==1.7.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
python3-saml==1.16.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
xmlsec==1.3.14
